chore: Apply feedback, move DeeplinkAction inside DeeplinkType
