--- app/src/main/java/com/infomaniak/drive/data/models/deeplink/DeeplinkAction.kt
@@ -1,48 +1,0 @@
-/*
- * Infomaniak kDrive - Android
- * Copyright (C) 2026 Infomaniak Network SA
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package com.infomaniak.drive.data.models.deeplink
-
-import kotlinx.parcelize.Parcelize
-
-
-@Parcelize
-sealed interface DeeplinkAction : DeeplinkType {
-    val driveId: Int
-
-    data class Collaborate(override val driveId: Int, val uuid: String) : DeeplinkAction {
-        override val isHandled: Boolean
-            get() = false
-    }
-
-    data class Drive(
-        val userId: Int? = null,
-        override val driveId: Int,
-        val roleFolder: RoleFolder
-    ) : DeeplinkAction {
-        override val isHandled: Boolean
-            get() = roleFolder.isHandled
-    }
-
-    data class Office(override val driveId: Int, val fileId: Int) : DeeplinkAction
-
-    companion object {
-        @Throws(InvalidFormatting::class)
-        fun from(actionType: String, action: String): DeeplinkAction = ActionType.from(actionType).build(action)
-
-    }
-}
